Edit key_paper loading
By using a list for key_papers, the processing can be done for all. At
the moment I am saving the complete list (which works for the key
papers) but it could be saved into separate rds files. The saving in
one list does not work for the results from the snownball search.